eventum_auth/requirements.txt
Django==2.0.3
djangorestframework==3.8
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt
gunicorn==19.7.1
PyJWT==1.6.1
pytz==2018.3
psycopg2>=2.7,<3.0
python-ldap
django-auth-ldap
django-ldapdb
